requirements.txt
# requirements.txt
# Confluence Q&A System with AutoGen and Azure Services

# AutoGen Framework
pyautogen>=0.2.0

# Azure Services
azure-identity>=1.15.0
azure-search-documents>=11.4.0
azure-cosmos>=4.5.1
azure-storage-blob>=12.19.0
# azure-openai>=1.0.0   

# Azure Functions
azure-functions==1.19.0

# OpenAI
openai>=1.0.0

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Graph Database
gremlinpython>=3.7.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
networkx>=3.2.0
anytree>=2.9.0

# Async Support
asyncio>=3.4.3
aiofiles>=23.2.1

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
python-dateutil>=2.8.2

# Logging and Monitoring
python-json-logger>=2.0.7

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

azure-functions
azure-storage-blob
requests

# Azure Cosmos DB Gremlin support
gremlinpython==3.6.2

# Azure Storage
azure-storage-blob==12.19.0

# Data processing
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0 